version 1.0.0: package interface changed + shipping source + dev server mode + build script + updated docs + exposing internal utilities
