+++ index.html
@@ -16,6 +16,8 @@
 
     <script>
 
+      XHRShaper.useGlobal();
+
       var loaded, total;
 
       var xhr = new XMLHttpRequest();
